Provide JSON codecs in client and server JsonSchemaEntities interpreters
Previously, we had a `JsonSchemaEntities` interpreter for each combination of client/server (Akka HTTP, Play framework, XMLHttpRequest) and JSON library (Play JSON, Circe). We now implement the `JsonSchemaEntities` interpreter directly in the client/server family, by hard-coding the usage of the `ujson.JsonSchemas` interpreter.

To simplify things even further, the client and server interpreters now automatically depend on the `openapi` interpreter, so that no additional dependency is required to use it.

Fixes #399